Added source files.
The source files contain almost the entirety of the Creepypasta Files app.
There are a few things that would need to be implemented (or
alternatively, removed) for a clean build. So as of this commit,
the application isn't automatically buildable when the repository is
cloned. This will need to be addressed.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>